+++ packages/terra-button-group/src/ButtonGroup.jsx
@@ -75,6 +75,7 @@ class ButtonGroup extends React.Component {
       cx(
         'button-group',
         { 'is-block': isBlock },
+        { 'is-touch-device': ('ontouchstart' in window) },
         theme.className,
       ),
       customProps.className,