+++ packages/terra-button/src/Button.jsx
@@ -245,6 +245,7 @@ class Button extends React.Component {
         { compact: isCompact },
         { 'is-active': this.state.active && !isDisabled },
         { 'is-focused': this.state.focused && !isDisabled },
+        { 'is-touch-device': ('ontouchstart' in window) },
         theme.className,
       ]),
       customProps.className,