+++ packages/terra-form-textarea/src/terra-dev-site/doc/example/IncompleteTextareaField.jsx
@@ -10,7 +10,6 @@ const DefaultTextAreaField = () => (
     isIncomplete
     inputAttrs={{
       name: 'requiredTextarea',
-      placeholder: 'Textarea placeholder',
     }}
   />
 );