+++ packages/terra-docs/tests/wdio/docs-spec.js
@@ -1,0 +1,9 @@
+// Terra.describeViewports('Docs', ['tiny', 'medium', 'large'], () => {
+//   describe('Default', () => {
+//     before(() => browser.url('/raw/tests/terra-docs/docs/default-docs'));
+
+//     it('validates the element', () => {
+//       Terra.validates.element();
+//     });
+//   });
+// });